Created MainView.jsx component to always render AsideMenu.jsx and conditionally other components.
