+++ src/components/AsideMenu/AsideMenu.jsx
@@ -1,23 +1,14 @@
-import React, {useState} from 'react';
+import React from 'react';
 import Tabs from "../Tabs/Tabs.jsx";
-import AddProject from "../AddProject/AddProject.jsx";
 
-export const AsideMenu = ({title}) => {
-    const [addProjectIsOpen, setAddProjectIsOpen] = useState(false);
-
-    const handleClick = () => {
-        setAddProjectIsOpen(isOpen => !isOpen);
-    }
+export const AsideMenu = ({title, onClick}) => {
 
     return (
-        <div className="flex w-full h-screen">
             <div className="aside-menu-container">
                 <h1 className="aside-menu-h1">{title}</h1>
-                <button className="aside-menu-button" onClick={handleClick}>+ Add Project</button>
+                <button className="aside-menu-button" onClick={onClick}>+ Add Project</button>
                 <Tabs className="aside-menu-tabs"/>
             </div>
-            {addProjectIsOpen && <AddProject/>}
-        </div>
     );
 };
